+++ NEWS.md
@@ -2,6 +2,11 @@
 # easy-samba news
 Stay updated on `easy-samba` development.
 
+### (2019-04-15 UTC) easy-samba 1.0.1 released
+Stable version 1.0.1 of `easy-samba` has been released.
+
+Click [here](https://github.com/adevur/docker-easy-samba/blob/master/docs/CHANGELOG.md#stable-101-2019-04-15-utc) for changelog.
+
 ### (2019-04-14 UTC) easy-samba 1.0.0 released
 Stable version 1.0.0 of `easy-samba` has been released.