+++ docs/CHANGELOG.md
@@ -2,11 +2,28 @@
 # easy-samba changelog
 Version history and changelogs of `adevur/easy-samba` docker image.
 
-### Current stable release: `1.0.0`
+### Current stable release: `1.0.1`
 
 ### Current long-term release: `no long-term release yet`
 
 ## Version history
+
+### [STABLE] 1.0.1 (2019-04-15 UTC)
+- New features:
+
+  - N/A
+
+- Bug fixes:
+
+  - Added forward-compatibility for future versions of `easy-samba`.
+  Now `easy-samba`'s logs will tell you if a `config.json` file is not compatible with the
+  version of `easy-samba` that you are currently using.
+  For example: `[ERROR] '/share/config.json' syntax is not correct: THIS CONFIGURATION FILE USES
+  FEATURES THAT REQUIRE EASY-SAMBA VERSION '1.1' OR NEWER.`
+
+- Security fixes:
+
+  - N/A
 
 ### [STABLE] 1.0.0 (2019-04-14 UTC)
 Initial release.